Implemented diffing and environment-based keys
You'll now need keys per environment. This will be dictated by
convention, but ideally it would be something like:

development.key
staging.key

And a file structure of:

decrypted/[project]/development/*.yml
decrypted/[project]/staging/*.yml

There is a diff log of every file that is updated that acts as a local
git history. This should be useful for resolving conflicts.